Rename the destination classes
