+++ appNekomp/src/commonMain/kotlin/com/chesire/nekomp/navigation/OriginScreen.kt
@@ -1,4 +1,4 @@
-package com.chesire.nekomp
+package com.chesire.nekomp.navigation
 
 enum class OriginScreen {
     Dashboard,
